fix(analyzer): address PR #8 review findings (security + correctness)
Eight distinct reviewer findings across Devin, Augment, Copilot, Kilo,
and Codex. Addressing all of them in a single commit.

SECURITY (high priority)

1. readFileTool path traversal (Devin P1, Augment P1, Codex P1, Kilo P1)
   The tool passed model-provided paths directly to readFile. Since the
   classifier runs over UNTRUSTED upstream commits, an attacker who
   controls upstream commit content can prompt-inject the model into
   reading /proc/self/environ, ../../etc/passwd, or .git/config, then
   exfiltrate secrets back through the tool output which is sent to
   external model providers. Concrete secret-exfiltration vector.

   Fix:
   - assertSafeRelativePath rejects absolute paths, null bytes, paths
     that resolve outside WORKSPACE_ROOT, and any .git/ segment.
   - Tool now reads `git show <ref>:<path>` so it reads the exact
     upstream commit state (addresses finding #5 below) rather than
     the fork's working tree.

2. Tools operate on fork's dev branch, not upstream commit state
   (Devin, medium)
   The old tools used `git grep` without specifying a ref, so they
   searched the fork's current HEAD. When classifying an upstream
   commit, the model saw fork-state callers, not upstream-state
   callers, making answers systematically wrong.

   Fix: Both tools now require a `ref` parameter (commit sha, tag, or
   branch). `read_file` uses `git show <ref>:<path>`. `grep_callers`
   uses `git grep ... <ref>`. Ref is validated against SAFE_REF_PATTERN
   (alphanumeric + ._/- only, max 200 chars) to prevent command
   injection via crafted refs.

3. grep_callers pattern injection (Augment medium, Copilot medium)
   `git grep` ran without -F, so symbols with regex metacharacters
   (foo.bar, a+b) matched unexpectedly. Symbols starting with - were
   also interpreted as flags.

   Fix: Use `git grep -F -e <symbol>` for fixed-string match, with
   `-e` ensuring the symbol is always data. Paths use `--` separator.

CORRECTNESS (medium priority)

4. APICallError instanceof check (Augment medium, Copilot)
   `err instanceof APICallError` misses errors from a different
   package instance (realm). Per AI SDK docs, use
   `APICallError.isInstance(err)`.

5. responseBody may be non-string (Augment medium, Copilot)
   `err.responseBody.slice()` and regex `.test()` would throw TypeError
   if the body was a Uint8Array or object, masking the original failure
   and preventing fallback.

   Fix: New stringifyBody() coerces via typeof check + JSON.stringify
   fallback before slicing/matching.

6. RateLimiter cache mismatch (Kilo)
   getLimiter() returned cached limiter without validating requested
   rate matches. Currently safe (fixed rates per provider) but a
   latent bug if callers ever request different values.

   Fix: Cache entries now include the requestsPerMinute value; a
   mismatched request throws a clear error rather than silently
   returning stale config.

7. slop_ratio_percent schema allowed floats (Copilot)
   Prompt said "integer 0-100", schema said z.number().min(0).max(100).
   Schema now enforces .int() too.

8. String.replace("_", "-") only fixes first underscore (Devin)
   HOLD_FOR_HUMAN became "hold-for_human" instead of "hold-for-human"
   for both issue labels and PR labels.

   Fix: Use /_/g regex for replace-all. (replaceAll would be cleaner
   but requires bumping TypeScript lib target past ES2021.)

HYGIENE (low priority)

9. ai + @ai-sdk/openai-compatible in main dependencies (Devin)
   Both packages are CI-only (used only by script/upstream-analyzer/).
   Moved to devDependencies so npm install doesn't ship them to end
   users of the plugin.

10. Workflow header comment outdated (Augment low)
    Referenced single-model gpt-4.1-mini / gpt-5-mini setup. Rewrote
    to document the new provider:model chain syntax and the three
    supported providers.

All LSP diagnostics clean. Typecheck clean. Actionlint clean.

+++ script/upstream-analyzer/schemas.ts
@@ -35,7 +35,7 @@ export const releaseSynthesisSchema = z.object({
     .describe("Overall merge recommendation"),
   confidence: z.enum(["high", "medium", "low"]),
   summary: z.string().min(1).max(2000).describe("3-5 sentence plain-English verdict"),
-  slop_ratio_percent: z.number().min(0).max(100),
+  slop_ratio_percent: z.number().int().min(0).max(100),
   breaking_changes: z.array(breakingChangeSchema),
   dependency_changes: z.array(z.string().min(1).max(300)),
   architecture_drift: z.array(z.string().min(1).max(300)),
